Decompiled func_80081E20

+++ include/code0/82480.h
@@ -6,7 +6,7 @@
 #include "static/119280.h"
 
 void func_800818A8(code0UnkStruct16 *, s16 id, s16 off, void *, void *, void *, void *);
-void func_80081E20(ProcPointer, ProcPointer, s32);
+void func_80081E20(ProcPointer, ProcPointer, s8);
 void func_80082410(Mtx *, f32, f32, f32);
 void func_80082448(void);
 void func_80083200(_119280UnkStruct1 *, s32);
